+++ R/FilterState.R
@@ -1662,11 +1662,34 @@ RangeFilterState <- R6::R6Class( # nolint
     # @return numeric(3) with names min, max, step - relevant for sliderInput
     get_pretty_range_inputs = function(values) {
       v_pretty_range <- pretty(values, n = 100)
+      min <- min(v_pretty_range)
+      max <- max(v_pretty_range)
+
+      step <- private$get_pretty_range_step(min, max, v_pretty_range)
+
       c(
-        min = v_pretty_range[1],
-        max = v_pretty_range[length(v_pretty_range)],
-        step = `if`(private$is_integer, 1L, v_pretty_range[2] - v_pretty_range[1])
+        min = min,
+        max = max,
+        step = step
       )
+    },
+    # @description gets pretty step size for range slider
+    #  adaptation of shiny's method (see shiny/R/input-slider.R function findStepSize)
+    # @param min (numeric) minimum of pretty values
+    # @param max (numeric) maximum of pretty values
+    # @param pretty_range (numeric(n)) vector of pretty values
+    # @return numeric(1) pretty step size for the sliderInput
+    get_pretty_range_step = function(min, max, pretty_range) {
+      range <- max - min
+
+      if (private$is_integer && range > 2) {
+        return(1L)
+      } else {
+        n_steps <- length(pretty_range) - 1
+        return(
+          signif(digits = 10, (max - min) / n_steps)
+        )
+      }
     },
     validate_selection = function(value) {
       if (!is.numeric(value)) {